create GenresList component
